test: drop the _more test-file names
